Added proper instruction cache

+++ Axh.Retro.CPU.X80.Contracts/Cache/IInstructionBlockCache.cs
@@ -5,12 +5,16 @@
     using Axh.Retro.CPU.X80.Contracts.Core;
     using Axh.Retro.CPU.X80.Contracts.Registers;
 
+    /// <summary>
+    /// Doesn't need to be thread safe
+    /// </summary>
+    /// <typeparam name="TRegisters"></typeparam>
     public interface IInstructionBlockCache<TRegisters> where TRegisters : IRegisters
     {
         Guid CacheId { get; }
 
         IInstructionBlock<TRegisters> GetOrSet(ushort address, Func<IInstructionBlock<TRegisters>> getInstanceFunc);
 
-        void InvalidateCache(ushort address, int length);
+        void InvalidateCache(ushort address, ushort length);
     }
 }
